# openapi/plystra.v0.0.1.yaml
components:
    schemas:
...
        ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIUser:
            properties:
                data:
                    $ref: '#/components/schemas/ApiOpenAPIUser'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIError:
            properties:
                audit_log_id:
                    example: audit_01hzyj6b6q8
                    type: string
                code:
                    example: VALIDATION_FAILED
                    type: string
                deny_code:
                    example: SCOPE_OUT_OF_BOUNDS
                    type: string
                details: {}
                message:
                    example: Request body is invalid JSON.
                    type: string
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
                trace_id:
                    example: trace_01hzyj6b6q8
                    type: string
            type: object
        ApiOpenAPIErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ApiOpenAPIError'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIUser:
            properties:
                data:
                    items:
                        $ref: '#/components/schemas/ApiOpenAPIUser'
                    nullable: true
                    type: array
                pagination:
                    $ref: '#/components/schemas/ApiOpenAPIPagination'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIPagination:
            properties:
                cursor:
                    nullable: true
                    type: string
                has_more:
                    example: false
                    type: boolean
                limit:
                    example: 50
                    type: integer
            type: object
...
        ApiOpenAPIUser:
            properties:
                created_at:
                    format: date-time
                    type: string
                deleted_at:
                    format: date-time
                    nullable: true
                    type: string
                email:
                    type: string
                id:
                    type: string
                last_login_at:
                    format: date-time
                    nullable: true
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                password_changed_at:
                    format: date-time
                    nullable: true
                    type: string
                phone:
                    type: string
                status:
                    type: string
                updated_at:
                    format: date-time
                    type: string
                username:
                    type: string
            type: object
...
        ApiUserMutationRequest:
            properties:
                actor:
                    $ref: '#/components/schemas/AuthzActorContext'
                audit_space_id:
                    type: string
                email:
                    type: string
                id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                password:
                    nullable: true
                    type: string
                phone:
                    nullable: true
                    type: string
                status:
                    nullable: true
                    type: string
                username:
                    nullable: true
                    type: string
            type: object
        AuthzActorContext:
            properties:
                binding_id:
                    type: string
                binding_status:
                    type: string
                member_display_name:
                    type: string
                member_id:
                    type: string
                member_status:
                    type: string
                relation_type:
                    type: string
                space_id:
                    type: string
                space_name:
                    type: string
                space_status:
                    type: string
                user_email:
                    type: string
                user_id:
                    type: string
                user_member_id:
                    type: string
                user_status:
                    type: string
            type: object
...
    securitySchemes:
        ApiKeyAuth:
            description: Scoped API key for server-to-server Core calls.
            in: header
            name: X-Plystra-API-Key
            type: apiKey
        BearerAuth:
            bearerFormat: opaque Plystra access token
            description: Use the access_token returned by POST /api/v1/auth/login or POST /api/v1/auth/refresh.
            scheme: bearer
            type: http
        MetricsTokenAuth:
            description: Dedicated metrics token when METRICS_TOKEN or PLYSTRA_METRICS_TOKEN is configured.
            in: header
            name: X-Plystra-Metrics-Token
            type: apiKey
info:
    description: Self-hosted Plystra Core API. Stable surfaces cover identity, authorization, resource registry, audit, admin control-plane, and scoped API keys; plugin, template, and Data Console routes are preview metadata surfaces.
    title: Plystra Core API
    version: 0.0.1
openapi: 3.0.3
paths:
...
    /api/v1/users:
        get:
            operationId: listUsers
            parameters:
                - description: Maximum number of rows to return.
                  in: query
                  name: limit
                  schema:
                    description: Maximum number of rows to return.
                    maximum: 200
                    minimum: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIUser'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List users
            tags:
                - Users
        post:
            operationId: createUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiUserMutationRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIUser'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Create a user
            tags:
                - Users
    /api/v1/users/{user_id}:
        get:
            operationId: getUser
            parameters:
                - in: path
                  name: user_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIUser'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Get a user
            tags:
                - Users
        patch:
            operationId: updateUser
            parameters:
                - in: path
                  name: user_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiUserMutationRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIUser'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Update a user
            tags:
                - Users
...
servers:
    - url: http://localhost:8080
...